+++ CHANGES.md
@@ -36,6 +36,10 @@
   - `eslint 9.21`
   - `storybook 8.6`
 
+* Added `selectedPlace` and `selectedPlaceGroup` to the list 
+  of available state properties. (#455)
+
+
 
 ## Changes in version 1.4.2